+++ .github/workflows/unit-test-coverage.yml
@@ -51,4 +51,4 @@ jobs:
 
       - name: Pytest coverage
         shell: bash
-        run: poetry run pytest tests --cov src --cov-report term-missing --cov-branch -v --cov-fail-under=80
+        run: poetry run pytest tests --cov=app --cov-report term-missing --cov-branch -v --cov-fail-under=80
